github.py
```python
from recruitment import RecruitmentProcess
from application import ApplicationForm
from database import Database
from utils import validateForm, saveApplicationToDatabase, retrieveApplicationFromDatabase

class GitHub:
    def __init__(self):
        self.recruitment = RecruitmentProcess()
        self.application_form = ApplicationForm()
        self.database = Database()

    def offerJob(self, user_id, job_id):
        job_position = self.recruitment.getJobPosition(job_id)
        salary = 180000
        benefits = ["Health insurance coverage", "Retirement savings plan", "Flexible work schedule", "Generous vacation and paid time off", "Professional development opportunities"]
        application_deadline = 24
        recruiter_contact = "GitHub Recruitment Team"
        return {"job_position": job_position, "salary": salary, "benefits": benefits, "application_deadline": application_deadline, "recruiter_contact": recruiter_contact}

    def startRecruitmentProcess(self, user_id, job_id):
        job_offer = self.offerJob(user_id, job_id)
        self.recruitment.sendJobOffer(user_id, job_offer)

    def submitApplication(self, user_id, job_id, application_form):
        if validateForm(application_form):
            saveApplicationToDatabase(self.database, user_id, job_id, application_form)
            self.recruitment.updateApplicationStatus(user_id, job_id, "Submitted")
        else:
            return "Invalid form"

    def retrieveApplication(self, user_id, job_id):
        application_status = retrieveApplicationFromDatabase(self.database, user_id, job_id)
        return application_status
```